Move all inline methods from header to source file
